backend/rag_processor.py
# backend/rag_processor.py
import os
from typing import List, Dict
import docx
import re

class RAGHRSystem:
    def __init__(self):
        pass
    
    def read_docx(self, file_path: str) -> str:
        """Чтение DOCX файла (заглушка)"""
        return "Текст вакансии или резюме"
    
    def prepare_documents(self, vacancy_text: str, resumes_texts: List[str]) -> List[Dict]:
        """Подготовка документов (заглушка)"""
        return []
    
    def create_vector_store(self, documents: List[Dict]):
        """Создание векторного хранилища (заглушка)"""
        pass
    
    def analyze_vacancy_match(self, vacancy_text: str, resume_text: str) -> str:
        """Анализ соответствия (заглушка)"""
        return """
        Анализ соответствия:
        Общее соответствие: 75%
        Совпадающие навыки: Python, SQL, Linux
        Отсутствующие навыки: Docker, Kubernetes
        """
    
    def extract_match_percentage(self, analysis_text: str) -> float:
        """Извлечение процента совпадения (заглушка)"""
        return 75.0
    
    def extract_skills(self, analysis_text: str) -> List[str]:
        """Извлечение навыков (заглушка)"""
        return ["Python", "SQL", "Linux"]
    
    def extract_missing_skills(self, analysis_text: str) -> List[str]:
        """Извлечение отсутствующих навыков (заглушка)"""
        return ["Docker", "Kubernetes"]