backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
python-docx==1.1.2
langchain-openai==0.1.2
spacy==3.7.4
scikit-learn==1.4.0
openai==1.12.0
